class_assignments/03_multiagents_system/pyproject.toml
[project]
name = "03-multiagents-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chainlit>=2.5.5",
    "python-dotenv>=1.1.0",
    "openai-agents>=0.0.17",
]
